lambda/port_deployment_trigger/requirements.txt
jsonpickle==2.2.0
requests==2.27.1
PyGithub==1.55
